Move threadpool stuff/team view down to RecordedRcon
